// client/src/Frontend/Components/CoreUI.tsx
export const Truncate = styled.div`
  ${({ maxWidth }: { maxWidth?: string }) => css`
    vertical-align: bottom;
    display: inline-block;
    white-space: nowrap;
    overflow: hidden;
    text-overflow: ellipsis;
    ${maxWidth !== undefined && `max-width: ${maxWidth};`};
  `}
`;

export const Title = styled(Truncate)`
  flex-grow: 1;
  text-align: left;
`;